+++ scripts/provision/debian/custom/basic/B_packages.sh
@@ -5,7 +5,18 @@ echo "Starting ${0}.."
 set -x
 
 update_dist=$ENV_basic_update_dist
-[[ -z "$update_dist" ]] && update_dist=false
+pkg_proxy_uri=$ENV_basic_pkg_proxy_uri
+
+if [[ -n "$pkg_proxy_uri" ]]; then
+cat <<EOF > /etc/apt/apt.conf.d/999_vagrant_proxy
+Acquire {
+  ftp::Proxy "${pkg_proxy_uri}";
+  http::Proxy "${pkg_proxy_uri}";
+  https::Proxy "${pkg_proxy_uri}";
+  ssh::Proxy "";
+}
+EOF
+fi
 
 apt-get update